+++ sdk/webpubsub/azure-messaging-webpubsubservice/azure/messaging/webpubsubservice/_web_pub_sub_service_client.py
@@ -7,43 +7,44 @@
 # --------------------------------------------------------------------------
 
 from copy import deepcopy
-from typing import Any, Awaitable, Optional, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
-from azure.core import AsyncPipelineClient
-from azure.core.rest import AsyncHttpResponse, HttpRequest
+from azure.core import PipelineClient
 from msrest import Deserializer, Serializer
 
-from ._configuration import WebpubsubServiceClientConfiguration
+from ._configuration import WebPubSubServiceClientConfiguration
 from .operations import HealthApiOperations, WebPubSubOperations
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
-    from typing import Dict
+    from typing import Any, Dict, Optional
 
-    from azure.core.credentials_async import AsyncTokenCredential
+    from azure.core.credentials import TokenCredential
+    from azure.core.rest import HttpRequest, HttpResponse
 
-class WebpubsubServiceClient:
-    """WebpubsubServiceClient.
+class WebPubSubServiceClient(object):
+    """WebPubSubServiceClient.
 
     :ivar health_api: HealthApiOperations operations
-    :vartype health_api: azure.messaging.webpubsubservice.aio.operations.HealthApiOperations
+    :vartype health_api: azure.messaging.webpubsubservice.operations.HealthApiOperations
     :ivar web_pub_sub: WebPubSubOperations operations
-    :vartype web_pub_sub: azure.messaging.webpubsubservice.aio.operations.WebPubSubOperations
+    :vartype web_pub_sub: azure.messaging.webpubsubservice.operations.WebPubSubOperations
     :param credential: Credential needed for the client to connect to Azure.
-    :type credential: ~azure.core.credentials_async.AsyncTokenCredential
+    :type credential: ~azure.core.credentials.TokenCredential
     :keyword endpoint: Service URL. Default value is ''.
     :paramtype endpoint: str
     """
 
     def __init__(
         self,
-        credential: "AsyncTokenCredential",
-        *,
-        endpoint: str = "",
-        **kwargs: Any
-    ) -> None:
-        self._config = WebpubsubServiceClientConfiguration(credential, **kwargs)
-        self._client = AsyncPipelineClient(base_url=endpoint, config=self._config, **kwargs)
+        credential,  # type: "TokenCredential"
+        **kwargs  # type: Any
+    ):
+        # type: (...) -> None
+        endpoint = kwargs.pop('endpoint', "")  # type: str
+
+        self._config = WebPubSubServiceClientConfiguration(credential, **kwargs)
+        self._client = PipelineClient(base_url=endpoint, config=self._config, **kwargs)
 
         self._serialize = Serializer()
         self._deserialize = Deserializer()
@@ -54,16 +55,17 @@ class WebpubsubServiceClient:
 
     def send_request(
         self,
-        request: HttpRequest,
-        **kwargs: Any
-    ) -> Awaitable[AsyncHttpResponse]:
+        request,  # type: HttpRequest
+        **kwargs  # type: Any
+    ):
+        # type: (...) -> HttpResponse
         """Runs the network request through the client's chained policies.
 
         >>> from azure.core.rest import HttpRequest
         >>> request = HttpRequest("GET", "https://www.example.org/")
         <HttpRequest [GET], url: 'https://www.example.org/'>
-        >>> response = await client.send_request(request)
-        <AsyncHttpResponse: 200 OK>
+        >>> response = client.send_request(request)
+        <HttpResponse: 200 OK>
 
         For more information on this code flow, see https://aka.ms/azsdk/python/protocol/quickstart
 
@@ -71,19 +73,22 @@ class WebpubsubServiceClient:
         :type request: ~azure.core.rest.HttpRequest
         :keyword bool stream: Whether the response payload will be streamed. Defaults to False.
         :return: The response of your network call. Does not do error handling on your response.
-        :rtype: ~azure.core.rest.AsyncHttpResponse
+        :rtype: ~azure.core.rest.HttpResponse
         """
 
         request_copy = deepcopy(request)
         request_copy.url = self._client.format_url(request_copy.url)
         return self._client.send_request(request_copy, **kwargs)
 
-    async def close(self) -> None:
-        await self._client.close()
+    def close(self):
+        # type: () -> None
+        self._client.close()
 
-    async def __aenter__(self) -> "WebpubsubServiceClient":
-        await self._client.__aenter__()
+    def __enter__(self):
+        # type: () -> WebPubSubServiceClient
+        self._client.__enter__()
         return self
 
-    async def __aexit__(self, *exc_details) -> None:
-        await self._client.__aexit__(*exc_details)
+    def __exit__(self, *exc_details):
+        # type: (Any) -> None
+        self._client.__exit__(*exc_details)