+++ sdk/webpubsub/azure-messaging-webpubsubservice/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 1.0.0b2 (Unreleased)
 
+- Add operations to client
+- Support AAD
 
 ## 1.0.0b1 (2021-04-27)
 
-Initial version
+- Initial version